Added scene handling

+++ src/game/actors/obstacle.cpp
@@ -3,13 +3,14 @@
 //
 
 #include "obstacle.h"
+#include "../gameManager.h"
 
-Obstacle::Obstacle(const char *path, int health, float x, float y) : Actor(path, health)
+Obstacle::Obstacle(const char *path, int health, int x, int y) : Actor(path, health)
 {
     sprite->setPosition(x, y);
 }
 
 void Obstacle::onDie()
 {
-
+    //GameManager::Instance
 }
